requirements.txt
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
decorator==4.4.1
Django==2.2.7
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
parso==0.5.1
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.5
pytest==5.3.0
pytest-django==3.7.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
zipp==0.6.0
